install mmpose, Keypoints note, context manager, predict, ota rename
