+++ mmyolo/models/task_modules/assigners/__init__.py
@@ -2,7 +2,7 @@
 from .batch_atss_assigner import BatchATSSAssigner
 from .batch_dsl_assigner import BatchDynamicSoftLabelAssigner
 from .batch_task_aligned_assigner import BatchTaskAlignedAssigner
-from .pose_sim_otaa_assigner import PoseSimOTAAssigner
+from .pose_sim_ota_assigner import PoseSimOTAAssigner
 from .utils import (select_candidates_in_gts, select_highest_overlaps,
                     yolov6_iou_calculator)
