+++ requirements/mmpose.txt
@@ -1,0 +1,1 @@
+mmpose>=1.0.0
