+++ requirements/tests.txt
@@ -5,6 +5,7 @@ isort==4.3.21
 kwarray
 memory_profiler
 mmcls>=1.0.0rc4
+mmpose>=1.0.0
 mmrazor>=1.0.0rc2
 mmrotate>=1.0.0rc1
 parameterized
